+++ CD/CD/Models/Mark.cs
@@ -6,7 +6,7 @@ namespace CD.Models
     {
         public Guid SubjectID { get; set; }
         public string MarkName { get; set; }
-        public int Result { get; set; }
+        public decimal Result { get; set; }
         public string Category { get; set; }
         public int Weight { get; set; }
     }
